I solved the problem with check-crime-prevented
Now I removed the check-crime function and just added the police-force-ID to a local variable in time-to-mobilise function which prints out the resource we get and the police force ID of that resource. I checked to confirm it was presenting the correct police force ID and this is the case.